added SHA-2 Hash Alg.

+++ examples/example_AES.cpp
@@ -12,10 +12,10 @@ int main()
     auto expected = vrock::utils::ByteArray::from_hex_string("c0749a1e19196e977506c0e2f1af6f4552");
     auto encrypted = vrock::security::encryption::AES::encrypt_gcm(data, key, iv, aad);
 
-    for (int i = 0; i < encrypted->length; ++i)
+    for (size_t i = 0; i < encrypted->length; ++i)
         std::cout << std::setw(2) << std::setfill('0') << std::hex << (int)encrypted->data[i] << ' ';
     std::cout << std::endl;
-    for (int i = 0; i < expected->length; ++i)
+    for (size_t i = 0; i < expected->length; ++i)
         std::cout << std::setw(2) << std::setfill('0') << std::hex << (int)expected->data[i] << ' ';
     auto decrypted = vrock::security::encryption::AES::decrypt_gcm(encrypted, key, iv, aad);
     std::cout << std::endl << decrypted->to_string() << std::endl;
